[testutil] Unify test certificate
