+++ ptlshs/example_test.go
@@ -11,6 +11,8 @@ import (
 	"net/http"
 	"strings"
 	"time"
+
+	"github.com/getlantern/tlsmasq/internal/testutil"
 )
 
 // Example demonstrates proxied TLS handshakes.
@@ -115,7 +117,7 @@ func Example() {
 
 // Each successful handshake will result in a signal sent on the input channel.
 func startTLSServer(successfulHandshakes chan<- struct{}) (addr string, err error) {
-	tlsListener, err := tls.Listen("tcp", "localhost:0", &tls.Config{Certificates: []tls.Certificate{cert}})
+	tlsListener, err := tls.Listen("tcp", "localhost:0", &tls.Config{Certificates: []tls.Certificate{testutil.Cert}})
 	if err != nil {
 		return "", err
 	}
@@ -143,33 +145,4 @@ func (l loggingTLSListener) Accept() (net.Conn, error) {
 	}
 	l.handshakes <- struct{}{}
 	return conn, nil
-}
-
-var (
-	certPem = []byte(`-----BEGIN CERTIFICATE-----
-MIIBhTCCASugAwIBAgIQIRi6zePL6mKjOipn+dNuaTAKBggqhkjOPQQDAjASMRAw
-DgYDVQQKEwdBY21lIENvMB4XDTE3MTAyMDE5NDMwNloXDTE4MTAyMDE5NDMwNlow
-EjEQMA4GA1UEChMHQWNtZSBDbzBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABD0d
-7VNhbWvZLWPuj/RtHFjvtJBEwOkhbN/BnnE8rnZR8+sbwnc/KhCk3FhnpHZnQz7B
-5aETbbIgmuvewdjvSBSjYzBhMA4GA1UdDwEB/wQEAwICpDATBgNVHSUEDDAKBggr
-BgEFBQcDATAPBgNVHRMBAf8EBTADAQH/MCkGA1UdEQQiMCCCDmxvY2FsaG9zdDo1
-NDUzgg4xMjcuMC4wLjE6NTQ1MzAKBggqhkjOPQQDAgNIADBFAiEA2zpJEPQyz6/l
-Wf86aX6PepsntZv2GYlA5UpabfT2EZICICpJ5h/iI+i341gBmLiAFQOyTDT+/wQc
-6MF9+Yw1Yy0t
------END CERTIFICATE-----`)
-	keyPem = []byte(`-----BEGIN EC PRIVATE KEY-----
-MHcCAQEEIIrYSSNQFaA2Hwf1duRSxKtLYX5CB04fSeQ6tF1aY/PuoAoGCCqGSM49
-AwEHoUQDQgAEPR3tU2Fta9ktY+6P9G0cWO+0kETA6SFs38GecTyudlHz6xvCdz8q
-EKTcWGekdmdDPsHloRNtsiCa697B2O9IFA==
------END EC PRIVATE KEY-----`)
-
-	cert tls.Certificate
-)
-
-func init() {
-	var err error
-	cert, err = tls.X509KeyPair(certPem, keyPem)
-	if err != nil {
-		panic(err)
-	}
 }
